requirements.txt
a2wsgi==1.10.8
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.1
apache-airflow-core==3.0.1
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.27.0
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-standard==1.1.0
apache-airflow-task-sdk==1.0.1
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.3.0
cadwyn==5.3.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
flask==2.2.5
fsspec==2025.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
importlib-metadata==8.4.0
iniconfig==2.1.0
itsdangerous==2.2.0
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
numpy==2.2.6
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.5.0
protobuf==4.25.7
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic-core==2.33.2
pygments==2.19.1
pyjwt==2.10.1
pytest==8.3.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.0
rich-toolkit==0.14.5
rpds-py==0.24.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==1.4.54
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
typer==0.15.3
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
zipp==3.21.0
